// client/src/types.ts
export interface Product {
  id: number;
  name: string;
  price: number;
  unit: string;
  seller: string;
  sellerUserId?: number;
  location: string;
  stock: number;
  image: string;
  category: string;
  badges?: string[];
  isVerified?: boolean;
  status?: string;
}
// ...
export interface LandingPageProps {}